RM#4520 : Code refactoring to a new service
