+++ axelor-sale/src/main/resources/i18n/messages_fr.csv
