+++ axelor-sale/src/main/resources/views/SaleOrderLine.xml
@@ -1053,7 +1053,8 @@
   <action-record
     name="action-sale-order-line-record-set-default-invoicing-mode-select"
     model="com.axelor.apps.sale.db.SaleOrderLine">
-    <field name="invoicingModeSelect" expr="eval:1" if="toInvoice"/>
+    <field name="invoicingModeSelect"
+      expr="eval:__repo__(SaleOrderLine).INVOICING_MODE_DIRECTLY" if="toInvoice"/>
     <field name="invoicingModeSelect" expr="eval:null" if="!toInvoice"/>
   </action-record>
